% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiViews.R
\name{GetPageViewsDay}
\alias{GetPageViewsDay}
\title{Wikipedia Page Views}
\usage{
GetPageViewsDay(page.title, date, language = "en")
}
\arguments{
\item{page.title}{Title of the page to get views}

\item{date}{get views of page on given date}

\item{language}{Language of wikipedia page}

\item{year}{get views of page for given year}
}
\value{
An list with date and views on that date
}
\description{
Get views per page. 5 function with day, month, year and from 2007 date intervals
}
\examples{
GetPageViewsDay("India",20151231)
GetPageViewsMonth(200712,"United_States","en")
GetPageViewsYear(2012,"China","zh")
GetPageViews("England","hi")
GetPageViewsNew("Hello_World","en")
GetPageViewsNew: gives the data from Jul 2015 - Present in a single request
}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\seealso{
\code{\link{}}
}
\keyword{Page}
\keyword{Stats}
\keyword{Wikipedia}
\keyword{view}

