% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiTop1000.R
\name{wiki_top1000_day}
\alias{wiki_top1000_day}
\title{Wikipedia top 1000 pages per day}
\usage{
wiki_top1000_day(date)
}
\arguments{
\item{date}{Date in format: YYYYMMDD}
}
\value{
An list with Json data
}
\description{
Get Top 1000 wikipedia pages for each Day
}
\examples{
wiki_top1000_day(20151231)

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\seealso{
\code{\link{}}
}
\keyword{1000}
\keyword{Top}
\keyword{pages}
\keyword{wikipedia}

