% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiTop1000.R
\name{wiki_top1000_month}
\alias{wiki_top1000_month}
\title{Wikipedia top 1000 pages per month}
\usage{
wiki_top1000_month(month.year)
}
\arguments{
\item{month.year}{Date in format: YYYYMM}
}
\value{
An list with Json data
}
\description{
Get Top 1000 wikipedia pages for Month(all Days of a month)
}
\examples{
wiki_top1000_month(201512)

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\seealso{
\code{\link{}}
}
\keyword{1000}
\keyword{Top}
\keyword{pages}
\keyword{wikipedia}

