% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiViews.R
\name{wiki_views}
\alias{wiki_views}
\title{Wikipedia Page Views}
\usage{
wiki_views(page.title, language = "en")
}
\arguments{
\item{page.title}{Title of the page to get views}

\item{language}{Language of wikipedia page}
}
\value{
An list with date and views on that date
}
\description{
Get views per page for wikipedia page from its creation date.
}
\note{
For faster data collection use wiki_views_new. With one restriction that 
data is only avalaible from 2015-07-01 with the new api API.
}
\examples{
wiki_views("England","hi")

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\seealso{
\code{\link{}}
}
\keyword{Page}
\keyword{Stats}
\keyword{Wikipedia}
\keyword{view}

