% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiViews.R
\name{wiki_views_day}
\alias{wiki_views_day}
\title{Wikipedia Page Views per day}
\usage{
wiki_views_day(page.title, date, language = "en")
}
\arguments{
\item{page.title}{Title of the page to get views}

\item{date}{get views of page on given date}

\item{language}{Language of wikipedia page}
}
\value{
An list with date and views on that date
}
\description{
Get views per page for a given day
}
\examples{
wiki_views_day("India",20151231)

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\seealso{
\code{\link{}}
}
\keyword{Page}
\keyword{Stats}
\keyword{Wikipedia}
\keyword{view}

