% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiViews.R
\name{wiki_views_new}
\alias{wiki_views_new}
\title{Wikipedia Page Views new ApI}
\usage{
wiki_views_new(page.title, language = "en")
}
\arguments{
\item{page.title}{Title of the page to get views}

\item{language}{Language of wikipedia page}
}
\value{
An list with date and views on that date
}
\description{
Get views per page from 20150107
}
\note{
This function is used to get data from 2015-07-01 to present in single request
}
\examples{
wiki_views_new("Hello_World","en")

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\seealso{
\code{\link{}}
}
\keyword{Page}
\keyword{Stats}
\keyword{Wikipedia}
\keyword{view}

