% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiViews.R
\name{wiki_views_year}
\alias{wiki_views_year}
\title{Wikipedia Page Views per year}
\usage{
wiki_views_year(page.title, year, language = "en")
}
\arguments{
\item{page.title}{Title of the page to get views}

\item{year}{get views of page for given year}

\item{language}{Language of wikipedia page}
}
\value{
An list with date and views on that date
}
\description{
Get views for wikipedia page based on given year.
}
\examples{
wiki_views_year("China",2012,"zh")

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\seealso{
\code{\link{}}
}
\keyword{Page}
\keyword{Stats}
\keyword{Wikipedia}
\keyword{view}

