% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiEdits.R
\name{wikiEdits}
\alias{wikiEdits}
\title{Wikipedia Page Edits}
\usage{
wikiEdits(page.title, language = "en")
}
\arguments{
\item{page.title}{Title of the page to get views}

\item{language}{Language of wikipedia page}
}
\value{
An list with number of edits of a page per month from creation of page
}
\description{
Get Number of edits of a wikipedia page. List of Ip Edits, Minor Edits and total Edits are returned
}
\details{
\code{language} The codes represent the language codes defined by ISO 639-1 and ISO 639-3, and
the decision of which language code to use is usually determined by the IETF language tag policy.
For more details please check List_of_Wikipedias page[1]
}
\note{
languageCodes is a dataframe with all language short codes
}
\examples{
wikiEdits("United_States","en")

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\references{
[1] https://en.wikipedia.org/wiki/List_of_Wikipedias
}
\seealso{
\code{\link{}}
}
\keyword{Edits}
\keyword{Page}
\keyword{Wikipedia}

