% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiLangCodes-Data.R
\docType{data}
\name{wikiLangCodes}
\alias{wikiLangCodes}
\title{Wikipedia languages and codes.}
\format{An data frame with 3 columns: Language, Language Local and Short Code.}
\source{
\href{#' https://en.wikipedia.org/wiki/List_of_Wikipedias}
}
\usage{
data(wikiLangCodes)
}
\description{
Wikipedia languages and codes.
}
\examples{
data(wikiLangCodes)
code <- wikiLangCodes[wikiLangCodes$Language == 'Japanese',3]
}
\references{
https://en.wikipedia.org/wiki/List_of_Wikipedias
}
\keyword{datasets}

