% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiTop1000.R
\name{WikiTop1000PerDay}
\alias{WikiTop1000PerDay}
\title{Wikipedia top 100 pages}
\usage{
WikiTop1000PerDay(date)
}
\arguments{
\item{date}{Date in format: YYYYMMDD}

\item{date.month.year}{Date in format: YYYYMM}
}
\value{
An list with Json data
}
\description{
Get Top 1000 wikipedia pages for Day / Month(all Days of a month)
}
\examples{
WikiTop1000PerDay(20151231)
WikiTop1000PerMonth(200712)

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\seealso{
\code{\link{}}
}
\keyword{1000}
\keyword{Top}
\keyword{pages}
\keyword{wikipedia}

